+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * When merging kapp config fragments, lists will always overwrite each other -- values will not be appended any more
 * The `action` name and value of `dry-run` will no longer be passed to kapps as variables
 * Outputs can now also take 'conditions'. They'll only be loaded if all conditions are true
+* Added a `kapps graph` command that only prints the graph of kapps that would be processed. It can also produce SVG visualisations of the graph.
 
 ## 0.9.0
 * Rename the 'cache' subcommand to 'workspace' because it wasn't clear that users should actually work inside cache directories (i.e. they're not temporary caches).